Safari-609.1.11.1/LayoutTests/platform/gtk/fast/canvas/canvas-createPattern-video-loading-expected.txt

Test the behavior of 2d canvas createPattern as a video is loaded

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('canvas').getContext('2d').createPattern(video, 'repeat') is null
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 194
PASS buffer[1] is within 6 of 193
PASS buffer[2] is within 6 of 14
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 17
PASS buffer[1] is within 6 of 197
PASS buffer[2] is within 6 of 195
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS buffer[0] is within 6 of 20
PASS buffer[1] is within 6 of 200
PASS buffer[2] is within 6 of 18
PASS successfullyParsed is true

TEST COMPLETE
